merge.py: Turn "too few nput files" error into a warning
There's not much wrong e.g. if in CI user requested just a single build
config, and it went thru the "merge" stage. Just generate "merged" report
for the single input config in this case, instead of erroring out. This
allows to avoid checking <2 condition in each and every CI job using
this tool.

Signed-off-by: Paul Sokolovsky <paul.sokolovsky@linaro.org>